+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.21] - 2024-10-29
+
+### Added
+
+- New nginx config
+- Add firewall & security messures
+- Bought a domain for shadowself
+- Bought a production server
+
+### Changed
+
+- Change db env variables
+- tmux.sh also connect to production server
+
 ## [0.0.20] - 2024-10-27
 
 ### Added